Update explanation as per comment from gbj.
This is for example 9; quoting gbj:

I'd say "working as designed and maybe an interesting pattern"

Hydration definitely assumes (and should assume) that the HTML rendered
on the server is what it's supposed to look like on first load. The
benefit of this is that we don't unnecessarily update every
potentially-reactive DOM node during hydration. The downside is that if
you intentionally render different things on the server and the client,
you'll see the server version. But then again this is actually an
upside, as it means there is no flash of server content followed by
client content.

The "maybe an interesting pattern" is that using an Effect to update it
is probably the correct way of doing it if this is what you want to do